Refactor GPTAsker class to include max_tokens parameter
The `GPTAsker` class in the `ask_gpt.py` script has been refactored to include a new parameter `max_tokens`. This parameter allows for dynamic control of the maximum number of tokens used in the API call, based on different conditions. The default value is 2000, but it can be increased to 5000 when OpenAI is supposed to edit or rewrite the article.

Also, modified copy_parsed_response output text, since only the enter key allows you to continue.